pyproject.toml
[project]
name = "sandbox-bench"
version = "0.1.0"
description = "Open source benchmark suite for AI agent sandbox providers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["sandbox", "benchmark", "ai-agents", "e2b", "modal", "daytona"]
authors = [
    { name = "Zach Wentz", email = "zkwentz@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
e2b = ["e2b>=0.17.0"]
daytona = ["daytona-sdk>=0.1.0"]
modal = ["modal>=0.60.0"]
codesandbox = ["codesandbox-sdk>=0.1.0"]
all = [
    "e2b>=0.17.0",
    "daytona-sdk>=0.1.0",
    "modal>=0.60.0",
    "codesandbox-sdk>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
sandbox-bench = "sandbox_bench.cli:main"

[project.urls]
Homepage = "https://github.com/zkwentz/sandbox-bench"
Repository = "https://github.com/zkwentz/sandbox-bench"
Issues = "https://github.com/zkwentz/sandbox-bench/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sandbox_bench"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
